tests/wpt/mozilla/tests/mozilla/mozbrowser/mozbrowser_loadevents.html

<head>
  <title>mozbrowserloadstart, mozbrowserconnected and mozbrowserloadend are dispatched</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>

    async_test(function(t) {

      var has_reloaded = false;
      var expectedEvents = [
        "mozbrowserloadstart",
        "mozbrowserconnected",
        "mozbrowserloadend",
        "mozbrowserloadstart",
        "mozbrowserconnected",
        "mozbrowserloadend",
      ];

      var receivedEvents = [];

      var iframe = document.createElement("iframe");
      iframe.mozbrowser = "true";
      iframe.src = "http://web-platform.test:8000";

      for (var event_name of ["mozbrowserloadstart",
                              "mozbrowserconnected",
                              "mozbrowserloadend"]) {
        iframe.addEventListener(event_name, t.step_func(e => {
          receivedEvents.push(e.type);
          if (receivedEvents.length == expectedEvents.length) {
            assert_array_equals(receivedEvents, expectedEvents);
            t.done();
          } else {
            if (e.type == "mozbrowserloadend") {
              iframe.reload();
            }
          }
        }));
      }

      document.body.appendChild(iframe);
    });

  </script>
</body>
